requirements.txt
numpy==1.23.5
opencv-python==4.8.1.78
paddleocr==2.7.0
scipy==1.11.4
scikit-image==0.21.0
pyyaml==6.0.1
tqdm==4.66.1
loguru==0.7.2
SQLAlchemy==2.0.23
pytesseract==0.3.10
lmdb==1.4.1
more-itertools==10.1.0
joblib==1.3.2
threadpoolctl==3.2.0
decorator==5.1.1
